// src/app/core/constants/routing.enum.ts
export enum ROUTING {
  HOME = '/',
  DASHBOARD = 'dashboard',
  PURCHASE_AGREEMENT = 'purchase-agreement',
  SALES_AGREEMENT = 'sales-agreement',
  VEHICLE_MASTER = 'vehicle-master',
  HP_APPLICATION = 'hp-application',
  FLOOR_STOCK_APPLICATION = 'floor-stock-application',
  INSURANCE_QUOTATION = 'insurance-quotation',
  CONTACT = 'contact',
}